+++ MainWindow.xaml.cs
@@ -193,7 +193,7 @@ public partial class MainWindow : Window
         ContentBox.ShowLineNumbers = conf.ShowLineNumbers;
         ContentBox.TextArea.Options.ShowSpaces = conf.ShowSpaces;
         ContentBox.TextArea.Options.ShowTabs = conf.ShowTabs;
-        ContentBox.TextArea.Options.ShowEndOfLine = conf.ShowLineFeed || conf.ShowCarriageReturn;
+        ContentBox.TextArea.Options.ShowEndOfLine = conf.ShowEndOfLine;
         ContentBox.HorizontalScrollBarVisibility = conf.WordWrap
             ? System.Windows.Controls.ScrollBarVisibility.Disabled
             : System.Windows.Controls.ScrollBarVisibility.Auto;